Made edits to HPForm for OnDamage saves
Made initiative increase the round counter at the top when it returns to the top of the turn order

Made a prompt request to make saving throws at the end of a creatures turn

Made a prompt request to make saving throws at the start of the next creature's turn, after the last creature's EoT save message is exited

Made duration of status effects count down at the end of the creature's turn (made Duration public to do this).

Made Status Effect Maker have Duration default to 1.